:art: Easy Popup > migrate to webpack
