+++ .npmignore
@@ -1,0 +1,11 @@
+.idea
+.gitignore
+.babelrc.json
+.eslintrc.json
+.prettierrc.json
+jsconfig.json
+/config
+/public
+/src
+/web
+/dev
