+++ dev/index.html
@@ -4,42 +4,12 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Easy Popup v0.0.5</title>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Nunito:wght@400;700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="src/easy-popup.css">
-    <style>
-        * {box-sizing:border-box;}
-        body {
-            background-color:#fff; color:#000; margin:0; min-height:100vh;
-            display:flex; align-items:center; justify-content:space-between; flex-direction:column;
-            font-family:"Nunito", sans-serif; font-size:18px;
-            line-height:1.5;
-        }
-        body:before {content:"";}
-        ::selection {
-            color:#fff;
-            background:cornflowerblue;
-        }
-        .container {
-            width:100%;
-            max-width:768px;
-            padding:0 20px;
-        }
-        a {color:cornflowerblue;}
-
-        footer {
-            padding:5vh 20px;
-            text-align:center;
-        }
-        footer a {text-decoration:none; display:inline-block;}
-    </style>
 </head>
 <body>
 
 
 <div class="container">
-    <h1>Easy Popup v0.0.5</h1>
+    <h1 data-title="">Easy Popup v0.0.5</h1>
     <p>Click here to <a href="#demo-1">open popup 1</a>.</p>
     <p>Or here
         <button data-easy-popup-toggle="demo-1">Open popup 1</button>
@@ -185,24 +155,5 @@
 
 </footer>
 
-<script src="src/easy-popup.js"></script>
-<script>
-    console.log(window.EasyPopupData);
-
-
-    EasyPopup.init('.demo-2', {
-        id: 'demo-2',
-        title: 'Demo 2',
-        triggerSelector: '.open-demo-2',
-        closeButtonHTML: '<span>Custom close button HTML</span>',
-        outerClass: 'outer-class-test',
-        theme: 'style-2'
-    });
-
-    EasyPopup.get('demo-1').open();
-
-
-    console.log(window.EasyPopupData);
-</script>
 </body>
 </html>
