+++ dev/style.scss
@@ -1,0 +1,59 @@
+/**
+ * Variables
+ */
+$font-size: 1rem;
+$font-family: 'Finlandica', sans-serif;
+$page-width: 768px;
+$color-primary: #f1c400;
+$color-on-primary: #141414;
+
+* {
+  box-sizing: border-box;
+  border: none;
+}
+
+body {
+  font-family: $font-family;
+  padding: 20px;
+  margin: 0;
+  background-color: #fff;
+  color: #000;
+  min-height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex-direction: column;
+  font-size: 18px;
+  line-height: 1.5;
+}
+
+body:before {
+  content: "";
+}
+
+::selection {
+  color: #fff;
+  background: cornflowerblue;
+}
+
+.container {
+  margin-left: auto;
+  margin-right: auto;
+  max-width: $page-width;
+  width: 100%;
+  padding: 0 20px;
+}
+
+a {
+  color: cornflowerblue;
+}
+
+footer {
+  padding: 5vh 20px;
+  text-align: center;
+}
+
+footer a {
+  text-decoration: none;
+  display: inline-block;
+}
