+++ package.json
@@ -1,13 +1,59 @@
 {
-  "devDependencies": {
-    "browser-sync": "^2.27.7",
-    "gulp": "^4.0.2",
-    "gulp-clean-css": "^4.3.0"
+  "name": "easy-popup",
+  "outputFilename": "easy-popup",
+  "prettyName": "Easy Popup",
+  "codeName": "EasyPopup",
+  "version": "0.0.6",
+  "description": "Super light-weight JavaScript library to create a simple popup with mobile layout.",
+  "homepage": "https://github.com/viivue/easy-popup",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:viivue/easy-popup"
   },
-  "dependencies": {
-    "gulp-rename": "^2.0.0",
-    "gulp-replace": "^1.1.3",
-    "gulp-uglify-es": "^3.0.0",
-    "readline-sync": "^1.4.10"
+  "author": {
+    "name": "phucbm",
+    "url": "https://github.com/phucbm"
+  },
+  "keywords": [
+    "phucbm",
+    "javascript",
+    "es6"
+  ],
+  "main": "./dist/easy-popup.module.js",
+  "license": "MIT",
+  "scripts": {
+    "dev": "webpack serve --config config/webpack.dev.js",
+    "build": "cross-env ENTRY=dev webpack --config config/webpack.build.js",
+    "sass:compressed": "sass src/_style.scss:dist/easy-select.css --no-source-map",
+    "optimize": "csso dist/easy-select.css --output dist/easy-select.min.css --stat",
+    "output-css": "npm run sass:compressed && npm run optimize",
+    "prod": "webpack --config config/webpack.prod.js && cross-env TARGET=module webpack --config config/webpack.prod.js && npm run output-css",
+    "publish": "npm run prod & npm publish"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.15.8",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/preset-env": "^7.15.8",
+    "babel-loader": "^8.2.2",
+    "babel-preset-es2015": "^6.24.1",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^9.1.0",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.4.0",
+    "css-minimizer-webpack-plugin": "^3.1.1",
+    "csso-cli": "^4.0.1",
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.3.2",
+    "mini-css-extract-plugin": "^2.4.2",
+    "postcss-loader": "^6.2.0",
+    "postcss-preset-env": "^6.7.0",
+    "sass": "^1.43.5",
+    "sass-loader": "^12.2.0",
+    "style-loader": "^3.3.0",
+    "terser-webpack-plugin": "^5.3.1",
+    "webpack": "^5.58.2",
+    "webpack-cli": "^4.9.0",
+    "webpack-dev-server": "^4.3.1",
+    "webpack-merge": "^5.8.0"
   }
 }
