+++ public/style/fonts.css
@@ -1,0 +1,38 @@
+/**
+ * Fonts
+ */
+@font-face {
+    font-family:"Finlandica";
+    src:url("../fonts/Finlandica-Bold.woff2") format("woff2"),
+    url("../fonts/Finlandica-Bold.woff") format("woff");
+    font-weight:bold;
+    font-style:normal;
+    font-display:swap;
+}
+
+@font-face {
+    font-family:"Finlandica";
+    src:url("../fonts/Finlandica-BoldItalic.woff2") format("woff2"),
+    url("../fonts/Finlandica-BoldItalic.woff") format("woff");
+    font-weight:bold;
+    font-style:italic;
+    font-display:swap;
+}
+
+@font-face {
+    font-family:"Finlandica";
+    src:url("../fonts/Finlandica-Italic.woff2") format("woff2"),
+    url("../fonts/Finlandica-Italic.woff") format("woff");
+    font-weight:normal;
+    font-style:italic;
+    font-display:swap;
+}
+
+@font-face {
+    font-family:"Finlandica";
+    src:url("../fonts/Finlandica-Regular.woff2") format("woff2"),
+    url("../fonts/Finlandica-Regular.woff") format("woff");
+    font-weight:normal;
+    font-style:normal;
+    font-display:swap;
+}
