+++ src/_style.scss
@@ -1,5 +1,5 @@
 /*!
- * Easy Popup v0.0.5
+ * Easy Popup v0.0.6
  * https://github.com/viivue/easy-popup
  */
 /* master */
